+++ module/chat/chat-processors.js
@@ -16,9 +16,7 @@ import { isNiceDiceEnabled, i18n, splitArgs, makeRegexPatternFrom } from '../../
 // import StatusProcessor from '../chat/status-processor.js'
 import StatusChatProcessor from '../chat/status.js'
 import SlamChatProcessor from '../chat/slam.js'
-import JB2AChatProcessor from '../chat/jb2a.js'
 import { Migration } from '../../lib/migration.js'
-
 
 export default function RegisterChatProcessors() {
   ChatProcessors.registerProcessor(new RollAgainstChatProcessor())
@@ -43,7 +41,6 @@ export default function RegisterChatProcessors() {
   ChatProcessors.registerProcessor(new SetEventFlagsChatProcessor())
   ChatProcessors.registerProcessor(new RemoteChatProcessor())
   ChatProcessors.registerProcessor(new SlamChatProcessor())
-  ChatProcessors.registerProcessor(new JB2AChatProcessor())
   ChatProcessors.registerProcessor(new LightChatProcessor())
   ChatProcessors.registerProcessor(new ForceMigrateChatProcessor())
   ChatProcessors.registerProcessor(new ReimportChatProcessor())
@@ -163,7 +160,7 @@ class ChatExecuteChatProcessor extends ChatProcessor {
     GURPS.chatreturn = false
     let args = splitArgs(line.substr(2))
     GURPS.chatargs = args
-    let m = Object.values(game.macros.entries).filter(m => m.name.startsWith(args[0]))
+    let m = Object.values(game.macros.contents).filter(m => m.name.startsWith(args[0]))
     if (m.length > 0) {
       this.send()
       m[0].execute()
@@ -281,7 +278,7 @@ class SelectChatProcessor extends ChatProcessor {
       // @self
       for (const a of game.actors.entities) {
         let users = a
-          .getUsers(CONST.ENTITY_PERMISSIONS.OWNER, true)
+          .getOwners()
           .filter(u => !u.isGM)
           .map(u => u.id)
         if (users.includes(game.user.id)) {
@@ -379,20 +376,18 @@ class UsesChatProcessor extends ChatProcessor {
       var eqt, key
       let m2 = m[3].trim().match(/^(o[\.:])?(.*)/i)
       let pattern = m2[2].trim()
-      if (!!pattern) 
-        [eqt, key] = actor.findEquipmentByName(pattern, !!m2[1])
+      if (!!pattern) [eqt, key] = actor.findEquipmentByName(pattern, !!m2[1])
       else if (this.msgs().event?.currentTarget) {
         pattern = '&lt;current equipment&gt;'
         let t = this.msgs().event?.currentTarget
         let k = $(t).closest('[data-key]').attr('data-key')
         // if we find a data-key, then we assume that we are on the character sheet, and if the target
-        // is equipment, apply to that equipment. 
+        // is equipment, apply to that equipment.
         if (!!k) {
           key = k
           eqt = getProperty(actor.data, key)
           // if its not equipment, ignore.
-          if (eqt.count == null)
-            eqt = null
+          if (eqt.count == null) eqt = null
         }
       }
       if (!eqt)
@@ -456,21 +451,19 @@ class QtyChatProcessor extends ChatProcessor {
       var eqt, key
       let m2 = m[2].trim().match(/^(o[\.:])?(.*)/i)
       let pattern = m2[2].trim()
-      if (!!pattern) 
-        [eqt, key] = actor.findEquipmentByName(pattern, !!m2[1])
+      if (!!pattern) [eqt, key] = actor.findEquipmentByName(pattern, !!m2[1])
       else if (this.msgs().event?.currentTarget) {
         pattern = '&lt;current equipment&gt;'
         let t = this.msgs().event?.currentTarget
         let k = $(t).closest('[data-key]').attr('data-key')
         // if we find a data-key, then we assume that we are on the character sheet, and if the target
-        // is equipment, apply to that equipment. 
+        // is equipment, apply to that equipment.
         if (!!k) {
           key = k
           eqt = getProperty(actor.data, key)
           // if its not equipment, ignore.
-          if (eqt.count == null)
-            eqt = null
-          }
+          if (eqt.count == null) eqt = null
+        }
       }
       if (!eqt)
         ui.notifications.warn(i18n('GURPS.chatNoEquipmentMatched', 'No equipment matched') + " '" + pattern + "'")
@@ -566,7 +559,7 @@ class TrackerChatProcessor extends ChatProcessor {
       } else if (!!m[5]) {
         if (max == 0) max = Number.MAX_SAFE_INTEGER
         let v = actor.data.data.additionalresources.tracker[tracker].value + delta
-         if (v > max) {
+        if (v > max) {
           ui.notifications.warn(
             `${i18n('GURPS.chatExceededMax', 'Exceeded MAX')}:${max} ${i18n('GURPS.for')} ${i18n(
               'GURPS.chatResourceTracker'
@@ -582,7 +575,7 @@ class TrackerChatProcessor extends ChatProcessor {
         }
         await actor.update({ ['data.additionalresources.tracker.' + tracker + '.value']: v })
         this.prnt(`${i18n('GURPS.chatResourceTracker')}${display} ${m[5]} = ${v}`)
-        answer = (v >= 0)
+        answer = v >= 0
       } else ui.notifications.warn(`${i18n('GURPS.chatUnrecognizedFormat', 'Unrecognized format')} '${line}'`)
     }
     return answer
@@ -590,11 +583,11 @@ class TrackerChatProcessor extends ChatProcessor {
 }
 
 class LightChatProcessor extends ChatProcessor {
-  help() { 
-    return '/li torch|t|off &lt;dim dist&gt; &lt;bright dist&gt; &lt;angle&gt;'
+  help() {
+    return '/li &lt;dim dist&gt; &lt;bright dist&gt; &lt;angle&gt; &lt;anim&gt;|off '
   }
   matches(line) {
-    this.match = line.match(/^\/(light|li) *(none|off)? *(\d+)? *(\d+)? *(\d+)? *(torch|t)? *(\d+)? *(\d+)?/i)
+    this.match = line.match(/^\/(light|li) *(none|off)? *(\d+)? *(\d+)? *(\d+)? *(#\w\w\w\w\w\w)? *(\w+)? *(\d+)? *(\d+)?/i)
     return !!this.match
   }
   async process(line) {
@@ -602,31 +595,40 @@ class LightChatProcessor extends ChatProcessor {
       ui.notifications.warn(i18n('GURPS.chatYouMustHaveACharacterSelected'))
       return
     }
-    let noAnimation = {"type": ""}
-    let torchAnimation = {"type": "torch", "speed": parseInt(this.match[7]) || 1, "intensity": parseInt(this.match[8]) || 1}
-    let data = {
-      "dimLight": null, 
-      "brightLight": null, 
-      "lightAngle": 360,
-      "lightAnimation": noAnimation
+    if (line.match(/^\/(light|li) *$/)) {
+      this.priv("Possible animations: " + Object.keys(CONFIG.Canvas.lightAnimations).join(', '))
+      return
     }
-  
-    if (!this.match[2]) {
-      data.dimLight = parseInt(this.match[3] || 3)
-      data.brightLight = parseInt(this.match[4] || 5)
-      data.lightAngle = parseInt(this.match[5] || 360)
-      if (!!this.match[6]) {
-        data.lightAnimation = torchAnimation
+    let type = this.match[7] || ''
+    if (!!type) {
+      let m = Object.keys(CONFIG.Canvas.lightAnimations).find(k => k.startsWith(type))
+      if (!m) {
+        ui.notifications.warn("Unknown light animation '" + type + "'.  Expected: " + Object.keys(CONFIG.Canvas.lightAnimations).join(', '))
+        return
       }
-      else
-        data.lightAnimation = noAnimation
+      type = m
     }
-    canvas.tokens.controlled.map(token => token.update(data));
+    let anim = { type: type, speed: parseInt(this.match[8]) || 1, intensity: parseInt(this.match[9]) || 1 }
+    let data = {
+      dimLight: 0,
+      brightLight: 0,
+      lightAngle: 360,
+      lightAnimation: anim,
+      '-=lightColor': null
+    }
+
+    if (!this.match[2]) {
+      if (this.match[6]) data.lightColor = this.match[6]
+      data.dimLight = parseInt(this.match[3] || 0)
+      data.brightLight = parseInt(this.match[4] || 0)
+      data.lightAngle = parseInt(this.match[5] || 360)
+    }
+    for (const t of canvas.tokens.controlled) await t.document.update(data)
   }
 }
 
 class ForceMigrateChatProcessor extends ChatProcessor {
- help() {
+  help() {
     return null
   }