+++ module/gurps.js
@@ -680,7 +680,7 @@ async function performAction(action, actor, event, targets) {
         targets
       )
       return true
-    } else ui.notifications.warn('You must have a character selected')
+    } else ui.notifications.warn(i18n('GURPS.chatYouMustHaveACharacterSelected'))
 
   if (action.type === 'derivedroll')
     if (!!actor) {
@@ -688,7 +688,7 @@ async function performAction(action, actor, event, targets) {
       formula = d6ify(df + action.formula)
       prefix = 'Rolling ' + action.derivedformula + action.formula + ' ' + action.desc
       if (!!action.costs) targetmods.push(GURPS.ModifierBucket.makeModifier(0, action.costs))
-    } else ui.notifications.warn('You must have a character selected')
+    } else ui.notifications.warn(i18n('GURPS.chatYouMustHaveACharacterSelected'))
 
   /*  let attr = action => {
     let target = action.target
@@ -997,11 +997,19 @@ GURPS.handleRoll = handleRoll
 // If the desc contains *Cost ?FP or *Max:9 then perform action
 async function applyModifierDesc(actor, desc) {
   if (!desc) return null
-  let parse = desc.replace(/.*\* ?[Cc]osts? (\d+) ?[Ff][Pp].*/g, '$1')
-  if (parse != desc && !!actor && !actor.isSelf) {
-    let fp = parseInt(parse)
-    fp = actor.data.data.FP.value - fp
-    await actor.update({ 'data.FP.value': fp })
+  let m = desc.match(/.*\* ?Costs? (\d+) ?([\w\(\)]+)/i)
+  if (!!m && !!actor && !actor.isSelf) {
+    let delta = parseInt(m[1])
+    let target = m[2]
+    if (target.match(/^[hf]p/i)) {
+      let k = target.toUpperCase()
+      delta = actor.data.data[k].value - delta
+      await actor.update({ ['data.' + k + '.value']: delta })
+    } 
+    if (target.match(/^tr/i)) {
+      await GURPS.ChatProcessors.startProcessingLines('/' + target + " -" + delta)
+      return null
+    }
   }
   parse = desc.replace(/.*\*[Mm]ax: ?(\d+).*/g, '$1')
   if (parse != desc) {
@@ -1320,7 +1328,7 @@ GURPS.whisperOtfToOwner = function (otf, overridetxt, event, blindcheck, actor) 
     if (overridetxt.includes('"')) overridetxt = "'" + overridetxt + "'"
     else overridetxt = '"' + overridetxt + '"'
   } else overridetxt = ''
-  let users = actor?.getUsers(CONST.ENTITY_PERMISSIONS.OWNER, true).filter(u => !u.isGM) || []
+  let users = actor?.getOwners().filter(u => !u.isGM) || []
   let botf = '[' + overridetxt + '!' + otf + ']'
   otf = '[' + overridetxt + otf + ']'
   let buttons = {}
@@ -1372,12 +1380,12 @@ GURPS.whisperOtfToOwner = function (otf, overridetxt, event, blindcheck, actor) 
 GURPS.sendOtfMessage = function (content, blindroll, users) {
   let msgData = {
     content: content,
-    user: game.user._id,
+    user: game.user.id,
     blind: blindroll,
   }
   if (!!users) {
     msgData.type = CONST.CHAT_MESSAGE_TYPES.WHISPER
-    msgData.whisper = users.map(it => it._id)
+    msgData.whisper = users.map(it => it.id)
   } else {
     msgData.type = CONST.CHAT_MESSAGE_TYPES.OOC
   }
@@ -1478,7 +1486,7 @@ Hooks.once('init', async function () {
   let src = 'systems/gurps/icons/gurps4e.png'
   if (game.i18n.lang == 'pt_br') src = 'systems/gurps/icons/gurps4e-pt_br.png'
   $('#logo').attr('src', src)
-
+  
   // set up all hitlocation tables (must be done before MB)
   HitLocation.init()
   DamageChat.initSettings()
@@ -1495,8 +1503,8 @@ Hooks.once('init', async function () {
   GURPS.ConditionalInjury = new GURPSConditionalInjury()
 
   // Define custom Entity classes
-  CONFIG.Actor.entityClass = GurpsActor
-  CONFIG.Item.entityClass = GurpsItem
+  CONFIG.Actor.documentClass = GurpsActor
+  CONFIG.Item.documentClass = GurpsItem
 
   // preload drag-and-drop image
   {
@@ -1553,14 +1561,15 @@ Hooks.once('ready', async function () {
   initializeDamageTables()
   ResourceTrackerManager.initSettings()
 
-  new ThreeD6({
-    popOut: false,
-    minimizable: false,
-    resizable: false,
-    id: 'ThreeD6',
-    template: 'systems/gurps/templates/threed6.html',
-    classes: [],
-  }).render(true)
+  if (game.settings.get(settings.SYSTEM_NAME, settings.SETTING_SHOW_3D6))
+    new ThreeD6({
+      popOut: false,
+      minimizable: false,
+      resizable: false,
+      id: 'ThreeD6',
+      template: 'systems/gurps/templates/threed6.html',
+      classes: [],
+    }).render(true)
 
   GURPS.currentVersion = SemanticVersion.fromString(game.system.data.version)
   // Test for migration
@@ -1661,7 +1670,7 @@ Hooks.once('ready', async function () {
         for (let i = 0; i < t.length; i++) {
           let el = t[i]
           let combatant = $(el).parents('.combatant').attr('data-combatant-id')
-          let target = game.combat.combatants.filter(c => c._id === combatant)[0]
+          let target = game.combat.combatants.filter(c => c.id === combatant)[0]
           if (!!target.actor?.data.data.additionalresources[$(el).attr('data-onethird')]) $(el).addClass('active')
         }
 
@@ -1674,7 +1683,7 @@ Hooks.once('ready', async function () {
             flag = true
           }
           let combatant = $(el).parents('.combatant').attr('data-combatant-id')
-          let target = game.combat.combatants.filter(c => c._id === combatant)[0]
+          let target = game.combat.combatants.filter(c => c.id === combatant)[0]
           target.actor.changeOneThirdStatus($(el).attr('data-onethird'), flag)
         })
       }
@@ -1686,7 +1695,7 @@ Hooks.once('ready', async function () {
         let elementMouseIsOver = document.elementFromPoint(ev.clientX, ev.clientY)
 
         let combatant = $(elementMouseIsOver).parents('.combatant').attr('data-combatant-id')
-        let target = game.combat.combatants.filter(c => c._id === combatant)[0]
+        let target = game.combat.combatants.filter(c => c.id === combatant)[0]
 
         let event = ev.originalEvent
         let dropData = JSON.parse(event.dataTransfer.getData('text/plain'))
@@ -1699,7 +1708,7 @@ Hooks.once('ready', async function () {
 
   game.socket.on('system.gurps', resp => {
     if (resp.type == 'updatebucket') {
-      if (resp.users.includes(game.user._id)) game.GURPS.ModifierBucket.updateModifierBucket(resp.bucket)
+      if (resp.users.includes(game.user.id)) game.GURPS.ModifierBucket.updateModifierBucket(resp.bucket)
     }
     if (resp.type == 'initiativeChanged') {
       CONFIG.Combat.initiative = {
@@ -1725,6 +1734,7 @@ Hooks.once('ready', async function () {
             let destEqt = getProperty(destactor.data, destKey)
             destactor.updateEqtCount(destKey, destEqt.count + resp.count)
           } else {
+            resp.itemData.data.equipped = true
             destactor.addNewItemData(resp.itemData)
           }
           game.socket.emit('system.gurps', {
@@ -1913,6 +1923,30 @@ Hooks.once('ready', async function () {
     }
     dragRuler.registerSystem('gurps', GURPSSpeedProvider)
   })
+  
+  // Translate attribute mappings if not in English
+  if (game.i18n.lang != 'en') {
+    console.log("Mapping " + game.i18n.lang + " translations into PARSELINK_MAPPINGS")
+    let mappings = {}
+    for (let k in GURPS.PARSELINK_MAPPINGS) {
+      let v = GURPS.PARSELINK_MAPPINGS[k]
+      let i = v.indexOf('.value')
+      let nk = v
+      if (i >= 0) {
+        nk = nk.substr(0, i)
+      }
+      nk = nk.replace(/\./g, '') // remove periods
+      nk = game.i18n.localize('GURPS.' + nk).toUpperCase()
+      if (!GURPS.PARSELINK_MAPPINGS[nk]) {
+        console.log(`Mapping '${k}' -> '${nk}'`)
+        mappings[nk] = GURPS.PARSELINK_MAPPINGS[k]
+      }
+    }
+    mappings = {...mappings, ...GURPS.PARSELINK_MAPPINGS}
+    GURPS.PARSELINK_MAPPINGS = mappings
+  }
+
+
 
   // End of system "READY" hook.
 })